+++ src/orchidrec/models/__init__.py
@@ -8,6 +8,7 @@ from orchidrec.models.io import load_model, model_from_state, save_model
 from orchidrec.models.item_knn import ItemKNN
 from orchidrec.models.popularity import Popularity
 from orchidrec.models.sequential_markov import SequentialMarkov
+from orchidrec.models.slim_elastic import SLIMElastic
 from orchidrec.models.user_knn import UserKNN
 
 __all__ = [
@@ -18,6 +19,7 @@ __all__ = [
     "ItemKNN",
     "Popularity",
     "Recommendation",
+    "SLIMElastic",
     "SequentialMarkov",
     "UserKNN",
     "load_model",